LayoutTests/media/track/track-load-from-element-readyState.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script>

            function trackLoaded()
            {
                consoleWrite("EVENT(load)");
                track = document.getElementById('testTrackSrc'); 
                testExpected("track.readyState", HTMLTrackElement.LOADED);
                endTest();
            }

        </script>
    </head>
    <body>
        <p>Tests the load event on HTMLTrackElement and LOADED readyState on TextTrack when src is set on the element.</p>
        <video>
            <track id="testTrackSrc" src="captions-webvtt/tc004-webvtt-file.vtt" onload="trackLoaded()" default>
	      </video>
    </body>
</html>
