LayoutTests/media/track/track-load-from-element-readyState-expected.txt

Tests the load event on HTMLTrackElement and LOADED readyState on TextTrack when src is set on the element.

EVENT(load)
EXPECTED (track.readyState == '2') OK
END OF TEST
